Factor out StaticThought component.
